Move topology/internal/checks to top level internal package
